pyproject.toml
[project]
name = "trees-everywhere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "uuid6>=2025.0.1",
]

[dependency-groups]
dev = ["ruff>=0.12.5"]

[tool.ruff]
line-length = 79
extend-exclude = [
    'manage.py',
    'trees_everywhere/settings.py',
    '*/admin.py',
    '*/migrations/',
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'ANN']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'
